Update absm
If the patch to the kernel in /proc/cmdline includes a prefix after "linux" and
before / we need to include it also in the path to the initramfs, as we assume that
they live in the same directory. This can happen for instance is /boot is in its own
partition with an ext4 file system, as in Fedora 37. This modify the lines beginning
with initrd in the boot entries for the snapshots.